Restructure the test project
This will be required for future test framework implementations
