--- GameServerTests/TestHelpers.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using LeagueSandbox.GameServer.Core.Logic;
-using LeagueSandbox.GameServer.Logic.Content;
-
-namespace GameServerTests
-{
-    public class TestHelpers
-    {
-
-    }
-}
